requirements.txt
pandas==2.2.3
selenium==4.27.1
patool==3.1.0
webdriver-manager==4.0.2